pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Minimal LangChain/LangGraph core scaffold with Bedrock and FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.121.2",
    "langchain==1.0.7",
    "langchain-aws==1.0.0",
    "langgraph==1.0.3",
    "langgraph-checkpoint-postgres>=2.0.21",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.2.8",
    "pydantic>=2.11.3",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.8.0",
]
